BUG: bug section present while no limitations (generateAppliDoc)
Bug section was supposed to be replaced by a comment if there were no
doclimitation in the app. It was based on the boolean value of the returned
value of `GetDocLimitations`. It seems that this method doesn't return a None
type but a string with value 'None'.

Now the section is replaced by a comment if the return value of
`GetDocLimitations` is None or 'None'.

--HG--
branch : manpages