Add the list of known contributors
These file contains the list of contributors who have signed the CLA
(Contributor License Agreement) or who are exempted from doing so
given their contractual commitment (according to the contract they or
their employer have signed with CNES, the French space agency holds
the copyright to their contributions).

When a new contributor signs the CLA, its identity must be added to
this file.